requirements.txt
streamlit
tensorflow
pillow
numpy
tensorflow.keras.models